+++ database/migrations/2018_03_27_110219_create_game_reward_table.php
@@ -18,8 +18,8 @@ class CreateGameRewardTable extends Migration
             $table->increments('id');
             $table->unsignedInteger('player_id')->comment('玩家id');
             $table->string('action')->comment('游戏行为');
-            $table->double('reward')->comment('赚取身价数');
-            $table->double('pet_exp')->comment('赚取宠物经验');
+            $table->decimal('reward', 18, 8)->comment('赚取身价数');
+            $table->decimal('pet_exp', 18, 8)->comment('赚取宠物经验');
             $table->timestamp('created_at')->default(DB::raw('CURRENT_TIMESTAMP'));
 
             $table->index('player_id');
